the vacation date

+++ study_monitor/monitor_clinet/src/main/java/com/example/monitor_clinet/entity/RuntimeDetail.java
@@ -10,8 +10,8 @@ public class RuntimeDetail {
     double cpuUsage;
     double memoryUsage;
     double diskUsage;
-    double networkUpload;
-    double networkDownload;
+    long networkUpload;
+    long networkDownload;
     double diskRead;
     double diskWrite;
 }
