+++ study_monitor/my-project-backend/src/main/java/com/example/entity/dto/RuntimeData.java
@@ -20,9 +20,9 @@ public class RuntimeData {
     @Column
     double diskUsage;
     @Column
-    double networkUpload;
+    long networkUpload;
     @Column
-    double networkDownload;
+    long networkDownload;
     @Column
     double diskRead;
     @Column
