+++ study_monitor/my-project-backend/src/main/java/com/example/entity/vo/request/RuntimeDetailVO.java
@@ -8,8 +8,8 @@ public class RuntimeDetailVO {
     double cpuUsage;
     double memoryUsage;
     double diskUsage;
-    double networkUpload;
-    double networkDownload;
+    long networkUpload;
+    long networkDownload;
     double diskRead;
     double diskWrite;
 }
